KamerTest.java: remove comment about code generator
I've experimented with delegating the responsibility of generating
a unique code to a `CodeGenerator` interface and implementation,
and then injecting `CodeGenerator` in a package-private constructor
for `Kamer` (`Kamer#Kamer(CodeGenerator)`), but this approach
results in separate constructors with repeated code and identical
behavior due to language constraints and the fact that the unique
code gets generated and assigned in the constructor. See below:

Kamer.java
```
// ...

	@Transient
	@Inject
	private CodeGenerator codeGenerator;

	{
		this.datum = LocalDateTime.now();
		this.deelnemers = new HashSet<>();
	}

	public Kamer() {
		this.kamerCode = codeGenerator.genereerCode(Kamer.KAMER_CODE_MAX_LENGTE);
	}

	Kamer(@NotNull CodeGenerator codeGenerator) {
		/*
		 * Not possible:
		 * this.codeGenerator = codeGenerator;
		 * this();
		 */

		this.kamerCode = codeGenerator.genereerCode(Kamer.KAMER_CODE_MAX_LENGTE);
	}

// ...
```

KamerTest.java
```
// ...
	@Test
	void zetCodeInConstructor() {
		CodeGenerator codeGenerator = new CodeGenerator() {
			@Override
			public String genereerCode(int maxLengte) {
				return "123456";
			}
		}

		kamer = new Kamer(codeGenerator);
		Assertions.assertEquals("123456", kamer.getKamerCode());
	}
// ...
```

This approach wouldn't actually properly test whether or not a unique code
gets assigned in the non-unit test constructor (`Kamer#Kamer()`), because
if the code in `Kamer#Kamer()` changes, the changes will have to manually
applied to `Kamer#Kamer(CodeGenerator)`, defeating the purpose of the
unit tests.