fix ConstraintViolationException when merging a `Kamer` instance
Discovered while Nathan Wijnberg <NAA.Wijnberg@student.han.nl> and I were testing
wether or not roles were persisted to the datastore. Nathan correctly guessed
that it might have something to do with `begeleider` in Kamer.java being
out-of-sync and thus no longer 'valid'.

To fix this, I've removed the `begeleider` field and corrected the JP relationships.
A side-effect of this is that `KamerRepository` is no longer required, since we no
longer need to manually save the `Begeleider` of a `Kamer` as this is done
automatically as a result of the JPA relationships.

Fixing this major bug took a long time, because the relevant tests needed to be
changed to meet our minimum branch coverage requirement of 80%.

The following has been changed:
* Deelnemer.java:
    * change the return type of `getKamer()` to `Kamer` and ensure it doesn't
      return `null`, as a `Deelnemer` is required to be associated to a `Kamer`.
    * add `setKamer()` (for internal use by Kamer.java)
    * add unit tests
* Kamer.java:
    * remove `begeleider` field (fixes ConstraintViolationException)
    * add `getBegeleider()`, similar to `Deelnemer#getKamer()`
    * add and refactor unit tests
* remove repository/
* move logic from `KamerRepository#maakKamer(String)` to AppService.java
* AppService.java:
    * update `kamer` after adding a `Deelnemer`
    * add and refactor unit tests