KamerService.java: check `kamerCode` against datastore when opening
Originally, we kept track of 'registered' rooms to potentially reduce
network traffic, however this failed to take into account existing
rooms or rooms that have been closed.

I've opted to load and check rooms from the datastore with the given
`kamerCode` as this solves the latter two problems. This still leaves
us with the first problem: network traffic, or more specifically: the
potential for a denial-of-service attack. This can be mitigated by
enforcing a maximum number of requests in a specific timeframe.

Implementing this seems out of scope, as normally a system admin
would take measures against denial-of-service attacks through
the use of `iptables`, for example.