Kamer.java: move code generation to CodeGenerator
After some thought, I decided it would be better to separate the
generation of codes from Kamer and delegate it to a CodeGenerator
class. This makes for conciser code since Kamer shouldn't care
about the code *pattern* it receives.

Furthermore, by utilizing the strategy pattern we can increase flexibility.
A use case for this would be when, in the future, a larger string is needed.