split {App,Kamer}Service.java into {App,Kamer}{Service,Controller}.java
By splitting each class into a service and a controller, we can remove
the dependency on a concrete implementation by utilizing an interface.
This improves flexibility and is in line with the dependency inversion
principle from SOLID.

As a result, `Event#voerUit(Deelnemer, Session)` has been changed into
`Event#voerUit(Deelnemer)` as the business logic shouldn't need to
know where the 'input' came from.

Lastly, this allows us to more easily switch to a different protocol
in the future if needed, as the business logic and application logic
is now handled through an interface.